+++ src/Sidekick.Business/Apis/Poe/Trade/Data/Stats/Pseudo/PseudoPattern.cs
@@ -5,7 +5,7 @@ namespace Sidekick.Business.Apis.Poe.Trade.Data.Stats.Pseudo
 {
     public class PseudoPattern
     {
-        public PseudoPattern(Regex regex, int multiplier = 1)
+        public PseudoPattern(Regex regex, double multiplier = 1)
         {
             Pattern = regex;
             Multiplier = multiplier;
@@ -13,7 +13,7 @@ namespace Sidekick.Business.Apis.Poe.Trade.Data.Stats.Pseudo
 
         public Regex Pattern { get; set; }
 
-        public int Multiplier { get; set; }
+        public double Multiplier { get; set; }
 
         public List<PseudoPatternMatch> Matches { get; set; } = new List<PseudoPatternMatch>();
